+++ aaaaaa/helper.py
@@ -1,13 +1,17 @@
 from __future__ import annotations
 
+import os
 from contextlib import contextmanager
 from copy import copy
 from typing import TYPE_CHECKING, Any, Union
+from unittest.mock import patch
 
 import torch
+from PIL import Image
+from typing_extensions import Protocol
 
 from modules import safe
-from modules.shared import opts
+from modules.shared import cmd_opts, opts
 
 if TYPE_CHECKING:
     # 타입 체커가 빨간 줄을 긋지 않게 하는 편법
@@ -49,17 +53,22 @@ def change_torch_load():
 
 
 @contextmanager
-def pause_total_tqdm():
-    orig = opts.data.get("multiple_tqdm", True)
-    try:
-        opts.data["multiple_tqdm"] = False
+def disable_safe_unpickle():
+    with (
+        patch.dict(os.environ, {"TORCH_FORCE_NO_WEIGHTS_ONLY_LOAD": "1"}, clear=False),
+        patch.object(cmd_opts, "disable_safe_unpickle", True),
+    ):
         yield
-    finally:
-        opts.data["multiple_tqdm"] = orig
 
 
 @contextmanager
-def preseve_prompts(p: PT):
+def pause_total_tqdm():
+    with patch.dict(opts.data, {"multiple_tqdm": False}, clear=False):
+        yield
+
+
+@contextmanager
+def preserve_prompts(p: PT):
     all_pt = copy(p.all_prompts)
     all_ng = copy(p.all_negative_prompts)
     try:
@@ -71,3 +80,7 @@ def preseve_prompts(p: PT):
 
 def copy_extra_params(extra_params: dict[str, Any]) -> dict[str, Any]:
     return {k: v for k, v in extra_params.items() if not callable(v)}
+
+
+class PPImage(Protocol):
+    image: Image.Image