+++ adetailer/common.py
@@ -10,7 +10,7 @@ from typing import Any, Generic, Optional, TypeVar
 
 from huggingface_hub import hf_hub_download
 from PIL import Image, ImageDraw
-from rich import print
+from rich import print  # noqa: A004  Shadowing built-in 'print'
 from torchvision.transforms.functional import to_pil_image
 
 REPO_ID = "Bingsu/adetailer"
@@ -22,6 +22,7 @@ T = TypeVar("T", int, float)
 class PredictOutput(Generic[T]):
     bboxes: list[list[T]] = field(default_factory=list)
     masks: list[Image.Image] = field(default_factory=list)
+    confidences: list[float] = field(default_factory=list)
     preview: Optional[Image.Image] = None
 
 
@@ -38,8 +39,9 @@ def hf_download(file: str, repo_id: str = REPO_ID, check_remote: bool = True) ->
     with suppress(Exception):
         return hf_hub_download(repo_id, file, local_files_only=True)
 
-    msg = f"[-] ADetailer: Failed to load model {file!r} from huggingface"
-    print(msg)
+    if check_remote:
+        msg = f"[-] ADetailer: Failed to load model {file!r} from huggingface"
+        print(msg)
     return "INVALID"
 
 
@@ -163,7 +165,7 @@ def create_bbox_from_mask(
     """
     bboxes = []
     for mask in masks:
-        mask = mask.resize(shape)
+        mask = mask.resize(shape)  # noqa: PLW2901
         bbox = mask.getbbox()
         if bbox is not None:
             bboxes.append(list(bbox))