+++ scripts/!adetailer.py
@@ -4,23 +4,24 @@ import platform
 import re
 import sys
 import traceback
+from collections.abc import Sequence
 from copy import copy
 from functools import partial
 from pathlib import Path
-from textwrap import dedent
 from typing import TYPE_CHECKING, Any, NamedTuple, cast
 
 import gradio as gr
 from PIL import Image, ImageChops
-from rich import print
+from rich import print  # noqa: A004  Shadowing built-in 'print'
 
 import modules
 from aaaaaa.conditional import create_binary_mask, schedulers
 from aaaaaa.helper import (
-    change_torch_load,
+    PPImage,
     copy_extra_params,
+    disable_safe_unpickle,
     pause_total_tqdm,
-    preseve_prompts,
+    preserve_prompts,
 )
 from aaaaaa.p_method import (
     get_i,
@@ -42,19 +43,22 @@ from adetailer import (
 from adetailer.args import (
     BBOX_SORTBY,
     BUILTIN_SCRIPT,
+    INPAINT_BBOX_MATCH_MODES,
     SCRIPT_DEFAULT,
     ADetailerArgs,
+    InpaintBBoxMatchMode,
     SkipImg2ImgOrig,
 )
 from adetailer.common import PredictOutput, ensure_pil_image, safe_mkdir
 from adetailer.mask import (
     filter_by_ratio,
-    filter_k_largest,
+    filter_k_by,
     has_intersection,
     is_all_black,
     mask_preprocess,
     sort_bboxes,
 )
+from adetailer.opts import dynamic_denoise_strength, optimal_crop_size
 from controlnet_ext import (
     CNHijackRestore,
     ControlNetExt,
@@ -177,25 +181,23 @@ class AfterDetailerScript(scripts.Script):
                 guidance_end=args.ad_controlnet_guidance_end,
             )
 
-    def is_ad_enabled(self, *args_) -> bool:
-        arg_list = [arg for arg in args_ if isinstance(arg, dict)]
-        if not args_ or not arg_list:
-            message = f"""
-                       [-] ADetailer: Invalid arguments passed to ADetailer.
-                           input: {args_!r}
-                           ADetailer disabled.
-                       """
-            print(dedent(message), file=sys.stderr)
+    def is_ad_enabled(self, *args) -> bool:
+        arg_list = [arg for arg in args if isinstance(arg, dict)]
+        if not arg_list:
             return False
 
-        ad_enabled = args_[0] if isinstance(args_[0], bool) else True
-        pydantic_args = []
+        ad_enabled = args[0] if isinstance(args[0], bool) else True
+
+        not_none = False
         for arg in arg_list:
             try:
-                pydantic_args.append(ADetailerArgs(**arg))
+                adarg = ADetailerArgs(**arg)
             except ValueError:  # noqa: PERF203
                 continue
-        not_none = not all(arg.need_skip() for arg in pydantic_args)
+            else:
+                if not adarg.need_skip():
+                    not_none = True
+                    break
         return ad_enabled and not_none
 
     def set_skip_img2img(self, p, *args_) -> None:
@@ -240,9 +242,6 @@ class AfterDetailerScript(scripts.Script):
         return False
 
     def get_args(self, p, *args_) -> list[ADetailerArgs]:
-        """
-        `args_` is at least 1 in length by `is_ad_enabled` immediately above
-        """
         args = [arg for arg in args_ if isinstance(arg, dict)]
 
         if not args:
@@ -252,21 +251,21 @@ class AfterDetailerScript(scripts.Script):
         if hasattr(p, "_ad_xyz"):
             args[0] = {**args[0], **p._ad_xyz}
 
-        all_inputs = []
+        all_inputs: list[ADetailerArgs] = []
 
         for n, arg_dict in enumerate(args, 1):
             try:
                 inp = ADetailerArgs(**arg_dict)
-            except ValueError as e:
-                msg = f"[-] ADetailer: ValidationError when validating {ordinal(n)} arguments"
-                if hasattr(e, "add_note"):
-                    e.add_note(msg)
-                else:
-                    print(msg, file=sys.stderr)
-                raise
+            except ValueError:
+                msg = f"[-] ADetailer: ValidationError when validating {ordinal(n)} arguments:"
+                print(msg, arg_dict, file=sys.stderr)
+                continue
 
             all_inputs.append(inp)
 
+        if not all_inputs:
+            msg = "[-] ADetailer: No valid arguments found."
+            raise ValueError(msg)
         return all_inputs
 
     def extra_params(self, arg_list: list[ADetailerArgs]) -> dict:
@@ -516,7 +515,10 @@ class AfterDetailerScript(scripts.Script):
 
     def get_sampler(self, p, args: ADetailerArgs) -> str:
         if args.ad_use_sampler:
+            if args.ad_sampler == "Use same sampler":
+                return p.sampler_name
             return args.ad_sampler
+
         if hasattr(p, "_ad_orig"):
             return p._ad_orig.sampler_name
         return p.sampler_name
@@ -524,7 +526,7 @@ class AfterDetailerScript(scripts.Script):
     def get_scheduler(self, p, args: ADetailerArgs) -> dict[str, str]:
         "webui >= 1.9.0"
         if not args.ad_use_sampler:
-            return {}
+            return {"scheduler": getattr(p, "scheduler", "Automatic")}
 
         if args.ad_scheduler == "Use same scheduler":
             value = getattr(p, "scheduler", "Automatic")
@@ -532,7 +534,7 @@ class AfterDetailerScript(scripts.Script):
             value = args.ad_scheduler
         return {"scheduler": value}
 
-    def get_override_settings(self, p, args: ADetailerArgs) -> dict[str, Any]:
+    def get_override_settings(self, _p, args: ADetailerArgs) -> dict[str, Any]:
         d = {}
 
         if args.ad_use_clip_skip:
@@ -553,7 +555,7 @@ class AfterDetailerScript(scripts.Script):
             d["sd_vae"] = args.ad_vae
         return d
 
-    def get_initial_noise_multiplier(self, p, args: ADetailerArgs) -> float | None:
+    def get_initial_noise_multiplier(self, _p, args: ADetailerArgs) -> float | None:
         return args.ad_noise_multiplier if args.ad_use_noise_multiplier else None
 
     @staticmethod
@@ -589,8 +591,8 @@ class AfterDetailerScript(scripts.Script):
         script_runner = copy(p.scripts)
         script_args = self.script_args_copy(p.script_args)
 
-        ad_only_seleted_scripts = opts.data.get("ad_only_seleted_scripts", True)
-        if not ad_only_seleted_scripts:
+        ad_only_selected_scripts = opts.data.get("ad_only_selected_scripts", True)
+        if not ad_only_selected_scripts:
             return script_runner, script_args
 
         ad_script_names_string: str = opts.data.get("ad_script_names", SCRIPT_DEFAULT)
@@ -614,20 +616,30 @@ class AfterDetailerScript(scripts.Script):
         script_runner.alwayson_scripts = filtered_alwayson
         return script_runner, script_args
 
-    def disable_controlnet_units(
-        self, script_args: list[Any] | tuple[Any, ...]
-    ) -> None:
-        for obj in script_args:
-            if "controlnet" in obj.__class__.__name__.lower():
-                if hasattr(obj, "enabled"):
-                    obj.enabled = False
-                if hasattr(obj, "input_mode"):
-                    obj.input_mode = getattr(obj.input_mode, "SIMPLE", "simple")
+    def disable_controlnet_units(self, script_args: Sequence[Any]) -> list[Any]:
+        new_args = []
+        for arg in script_args:
+            if "controlnet" in arg.__class__.__name__.lower():
+                new = copy(arg)
+                if hasattr(new, "enabled"):
+                    new.enabled = False
+                if hasattr(new, "input_mode"):
+                    new.input_mode = getattr(new.input_mode, "SIMPLE", "simple")
+                new_args.append(new)
 
-            elif isinstance(obj, dict) and "module" in obj:
-                obj["enabled"] = False
+            elif isinstance(arg, dict) and "module" in arg:
+                new = copy(arg)
+                new["enabled"] = False
+                new_args.append(new)
 
-    def get_i2i_p(self, p, args: ADetailerArgs, image):
+            else:
+                new_args.append(arg)
+
+        return new_args
+
+    def get_i2i_p(
+        self, p, args: ADetailerArgs, image: Image.Image
+    ) -> StableDiffusionProcessingImg2Img:
         seed, subseed = self.get_seed(p)
         width, height = self.get_width_height(p, args)
         steps = self.get_steps(p, args)
@@ -685,7 +697,7 @@ class AfterDetailerScript(scripts.Script):
         i2i._ad_inner = True
 
         if args.ad_controlnet_model != "Passthrough" and controlnet_type != "forge":
-            self.disable_controlnet_units(i2i.script_args)
+            i2i.script_args = self.disable_controlnet_units(i2i.script_args)
 
         if args.ad_controlnet_model not in ["None", "Passthrough"]:
             self.update_controlnet_args(i2i, args)
@@ -695,6 +707,9 @@ class AfterDetailerScript(scripts.Script):
         return i2i
 
     def save_image(self, p, image, *, condition: str, suffix: str) -> None:
+        if not opts.data.get(condition, False):
+            return
+
         i = get_i(p)
         if p.all_prompts:
             i %= len(p.all_prompts)
@@ -703,18 +718,22 @@ class AfterDetailerScript(scripts.Script):
             save_prompt = p.prompt
         seed, _ = self.get_seed(p)
 
-        if opts.data.get(condition, False):
-            images.save_image(
-                image=image,
-                path=p.outpath_samples,
-                basename="",
-                seed=seed,
-                prompt=save_prompt,
-                extension=opts.samples_format,
-                info=self.infotext(p),
-                p=p,
-                suffix=suffix,
-            )
+        ad_save_images_dir: str = opts.data.get("ad_save_images_dir", "")
+
+        if not ad_save_images_dir.strip():
+            ad_save_images_dir = p.outpath_samples
+
+        images.save_image(
+            image=image,
+            path=ad_save_images_dir,
+            basename="",
+            seed=seed,
+            prompt=save_prompt,
+            extension=opts.samples_format,
+            info=self.infotext(p),
+            p=p,
+            suffix=suffix,
+        )
 
     def get_ad_model(self, name: str):
         if name not in model_mapping:
@@ -731,7 +750,7 @@ class AfterDetailerScript(scripts.Script):
         pred = filter_by_ratio(
             pred, low=args.ad_mask_min_ratio, high=args.ad_mask_max_ratio
         )
-        pred = filter_k_largest(pred, k=args.ad_mask_k_largest)
+        pred = filter_k_by(pred, k=args.ad_mask_k, by=args.ad_mask_filter_method)
         pred = self.sort_bboxes(pred)
         masks = mask_preprocess(
             pred.masks,
@@ -758,26 +777,21 @@ class AfterDetailerScript(scripts.Script):
         i2i.negative_prompt = negative_prompt
 
     @staticmethod
-    def compare_prompt(p, extra_params: dict[str, Any], processed, n: int = 0):
-        if not hasattr(p, "_ad_extra_params_result"):
-            p._ad_extra_params_result = {}
-
+    def compare_prompt(extra_params: dict[str, Any], processed, n: int = 0):
         pt = "ADetailer prompt" + suffix(n)
         if pt in extra_params and extra_params[pt] != processed.all_prompts[0]:
             print(
                 f"[-] ADetailer: applied {ordinal(n + 1)} ad_prompt: {processed.all_prompts[0]!r}"
             )
-            p._ad_extra_params_result[pt] = processed.all_prompts[0]
 
         ng = "ADetailer negative prompt" + suffix(n)
         if ng in extra_params and extra_params[ng] != processed.all_negative_prompts[0]:
             print(
                 f"[-] ADetailer: applied {ordinal(n + 1)} ad_negative_prompt: {processed.all_negative_prompts[0]!r}"
             )
-            p._ad_extra_params_result[ng] = processed.all_negative_prompts[0]
 
     @staticmethod
-    def get_i2i_init_image(p, pp):
+    def get_i2i_init_image(p, pp: PPImage):
         if is_skip_img2img(p):
             return p.init_images[0]
         return pp.image
@@ -792,25 +806,108 @@ class AfterDetailerScript(scripts.Script):
         img2img_mask: Image.Image, ad_mask: list[Image.Image]
     ) -> list[Image.Image]:
         if ad_mask and img2img_mask.size != ad_mask[0].size:
-            img2img_mask = img2img_mask.resize(ad_mask[0].size, resample=images.LANCZOS)
+            img2img_mask = img2img_mask.resize(ad_mask[0].size, resample=Image.LANCZOS)
         return [mask for mask in ad_mask if has_intersection(img2img_mask, mask)]
 
     @staticmethod
     def get_image_mask(p) -> Image.Image:
         mask = p.image_mask
+        mask = ensure_pil_image(mask, "L")
         if getattr(p, "inpainting_mask_invert", False):
             mask = ImageChops.invert(mask)
         mask = create_binary_mask(mask)
 
-        if is_skip_img2img(p):
-            if hasattr(p, "init_images") and p.init_images:
-                width, height = p.init_images[0].size
-            else:
-                msg = "[-] ADetailer: no init_images."
-                raise RuntimeError(msg)
-        else:
-            width, height = p.width, p.height
+        width, height = p.width, p.height
+        if is_skip_img2img(p) and hasattr(p, "init_images") and p.init_images:
+            width, height = p.init_images[0].size
         return images.resize_image(p.resize_mode, mask, width, height)
+
+    @staticmethod
+    def get_dynamic_denoise_strength(
+        denoise_strength: float, bbox: Sequence[Any], image_size: tuple[int, int]
+    ):
+        denoise_power = opts.data.get("ad_dynamic_denoise_power", 0)
+        if denoise_power == 0:
+            return denoise_strength
+
+        modified_strength = dynamic_denoise_strength(
+            denoise_power=denoise_power,
+            denoise_strength=denoise_strength,
+            bbox=bbox,
+            image_size=image_size,
+        )
+
+        print(
+            f"[-] ADetailer: dynamic denoising -- {denoise_strength:.2f} -> {modified_strength:.2f}"
+        )
+
+        return modified_strength
+
+    @staticmethod
+    def get_optimal_crop_image_size(
+        inpaint_width: int, inpaint_height: int, bbox: Sequence[Any]
+    ) -> tuple[int, int]:
+        calculate_optimal_crop = opts.data.get(
+            "ad_match_inpaint_bbox_size", InpaintBBoxMatchMode.OFF.value
+        )
+
+        optimal_resolution: tuple[int, int] | None = None
+
+        # Off
+        if calculate_optimal_crop == InpaintBBoxMatchMode.OFF.value:
+            return (inpaint_width, inpaint_height)
+
+        # Strict (SDXL only)
+        if calculate_optimal_crop == InpaintBBoxMatchMode.STRICT.value:
+            if not shared.sd_model.is_sdxl:
+                msg = "[-] ADetailer: strict inpaint bounding box size matching is only available for SDXL. Use Free mode instead."
+                print(msg)
+                return (inpaint_width, inpaint_height)
+
+            optimal_resolution = optimal_crop_size.sdxl(
+                inpaint_width, inpaint_height, bbox
+            )
+
+        # Free
+        elif calculate_optimal_crop == InpaintBBoxMatchMode.FREE.value:
+            optimal_resolution = optimal_crop_size.free(
+                inpaint_width, inpaint_height, bbox
+            )
+
+        if optimal_resolution is None:
+            msg = "[-] ADetailer: unsupported inpaint bounding box match mode. Original inpainting dimensions will be used."
+            print(msg)
+            return (inpaint_width, inpaint_height)
+
+        # Only use optimal dimensions if they're different enough to current inpaint dimensions.
+        if (
+            abs(optimal_resolution[0] - inpaint_width) > inpaint_width * 0.1
+            or abs(optimal_resolution[1] - inpaint_height) > inpaint_height * 0.1
+        ):
+            print(
+                f"[-] ADetailer: inpaint dimensions optimized -- {inpaint_width}x{inpaint_height} -> {optimal_resolution[0]}x{optimal_resolution[1]}"
+            )
+
+        return optimal_resolution
+
+    def fix_p2(  # noqa: PLR0913
+        self, p, p2, pp: PPImage, args: ADetailerArgs, pred: PredictOutput, j: int
+    ):
+        seed, subseed = self.get_seed(p)
+        p2.seed = self.get_each_tab_seed(seed, j)
+        p2.subseed = self.get_each_tab_seed(subseed, j)
+        p2.denoising_strength = self.get_dynamic_denoise_strength(
+            p2.denoising_strength, pred.bboxes[j], pp.image.size
+        )
+
+        p2.cached_c = [None, None]
+        p2.cached_uc = [None, None]
+
+        # Don't override user-defined dimensions.
+        if not args.ad_use_inpaint_width_height:
+            p2.width, p2.height = self.get_optimal_crop_image_size(
+                p2.width, p2.height, pred.bboxes[j]
+            )
 
     @rich_traceback
     def process(self, p, *args_):
@@ -864,7 +961,7 @@ class AfterDetailerScript(scripts.Script):
         p._ad_is_hr_pass = True
 
     def _postprocess_image_inner(
-        self, p, pp, args: ADetailerArgs, *, n: int = 0
+        self, p, pp: PPImage, args: ADetailerArgs, *, n: int = 0
     ) -> bool:
         """
         Returns
@@ -879,23 +976,23 @@ class AfterDetailerScript(scripts.Script):
         i = get_i(p)
 
         i2i = self.get_i2i_p(p, args, pp.image)
-        seed, subseed = self.get_seed(p)
         ad_prompts, ad_negatives = self.get_prompt(p, args)
 
         is_mediapipe = args.is_mediapipe()
 
-        kwargs = {}
         if is_mediapipe:
-            predictor = mediapipe_predict
-            ad_model = args.ad_model
-        else:
-            predictor = ultralytics_predict
-            ad_model = self.get_ad_model(args.ad_model)
-            kwargs["device"] = self.ultralytics_device
-            kwargs["classes"] = args.ad_model_classes
+            pred = mediapipe_predict(args.ad_model, pp.image, args.ad_confidence)
 
-        with change_torch_load():
-            pred = predictor(ad_model, pp.image, args.ad_confidence, **kwargs)
+        else:
+            ad_model = self.get_ad_model(args.ad_model)
+            with disable_safe_unpickle():
+                pred = ultralytics_predict(
+                    ad_model,
+                    image=pp.image,
+                    confidence=args.ad_confidence,
+                    device=self.ultralytics_device,
+                    classes=args.ad_model_classes,
+                )
 
         if pred.preview is None:
             print(
@@ -944,8 +1041,6 @@ class AfterDetailerScript(scripts.Script):
             p2.seed = self.get_each_tab_seed(seed, j)
             p2.subseed = self.get_each_tab_seed(subseed, j)
 
-            p2.cached_c = [None, None]
-            p2.cached_uc = [None, None]
             try:
                 processed = process_images(p2)
             except NansException as e:
@@ -955,7 +1050,11 @@ class AfterDetailerScript(scripts.Script):
             finally:
                 p2.close()
 
-            self.compare_prompt(p, p.extra_generation_params, processed, n=n)
+            if not processed.images:
+                processed = None
+                break
+
+            self.compare_prompt(p.extra_generation_params, processed, n=n)
             p2 = copy(i2i)
             p2.init_images = [processed.images[0]]
 
@@ -966,7 +1065,7 @@ class AfterDetailerScript(scripts.Script):
         return False
 
     @rich_traceback
-    def postprocess_image(self, p, pp, *args_):
+    def postprocess_image(self, p, pp: PPImage, *args_):
         if getattr(p, "_ad_disabled", False) or not self.is_ad_enabled(*args_):
             return
 
@@ -978,7 +1077,7 @@ class AfterDetailerScript(scripts.Script):
 
         if need_call_postprocess(p):
             dummy = Processed(p, [], p.seed, "")
-            with preseve_prompts(p):
+            with preserve_prompts(p):
                 p.scripts.postprocess(copy(p), dummy)
 
         is_processed = False
@@ -996,16 +1095,12 @@ class AfterDetailerScript(scripts.Script):
             )
 
         if need_call_process(p):
-            with preseve_prompts(p):
+            with preserve_prompts(p):
                 copy_p = copy(p)
-                if hasattr(p.scripts, "before_process"):
-                    p.scripts.before_process(copy_p)
+                p.scripts.before_process(copy_p)
                 p.scripts.process(copy_p)
 
         self.write_params_txt(params_txt_content)
-
-        if hasattr(p, "_ad_extra_params_result"):
-            p.extra_generation_params.update(p._ad_extra_params_result)
 
 
 def on_after_component(component, **_kwargs):
@@ -1035,7 +1130,7 @@ def on_ui_settings():
         "ad_extra_models_dir",
         shared.OptionInfo(
             default="",
-            label="Extra paths to scan adetailer models seperated by vertical bars(|)",
+            label="Extra paths to scan adetailer models separated by vertical bars(|)",
             component=gr.Textbox,
             section=section,
         )
@@ -1044,19 +1139,33 @@ def on_ui_settings():
     )
 
     shared.opts.add_option(
+        "ad_save_images_dir",
+        shared.OptionInfo(
+            default="",
+            label="Output directory for adetailer images",
+            component=gr.Textbox,
+            section=section,
+        ),
+    )
+
+    shared.opts.add_option(
         "ad_save_previews",
-        shared.OptionInfo(False, "Save mask previews", section=section),
+        shared.OptionInfo(default=False, label="Save mask previews", section=section),
     )
 
     shared.opts.add_option(
         "ad_save_images_before",
-        shared.OptionInfo(False, "Save images before ADetailer", section=section),
+        shared.OptionInfo(
+            default=False, label="Save images before ADetailer", section=section
+        ),
     )
 
     shared.opts.add_option(
-        "ad_only_seleted_scripts",
+        "ad_only_selected_scripts",
         shared.OptionInfo(
-            True, "Apply only selected scripts to ADetailer", section=section
+            default=True,
+            label="Apply only selected scripts to ADetailer",
+            section=section,
         ),
     )
 
@@ -1090,7 +1199,35 @@ def on_ui_settings():
     shared.opts.add_option(
         "ad_same_seed_for_each_tab",
         shared.OptionInfo(
-            False, "Use same seed for each tab in adetailer", section=section
+            default=False,
+            label="Use same seed for each tab in adetailer",
+            section=section,
+        ),
+    )
+
+    shared.opts.add_option(
+        "ad_dynamic_denoise_power",
+        shared.OptionInfo(
+            default=0,
+            label="Power scaling for dynamic denoise strength based on bounding box size",
+            component=gr.Slider,
+            component_args={"minimum": -10, "maximum": 10, "step": 0.01},
+            section=section,
+        ).info(
+            "Smaller areas get higher denoising, larger areas less. Maximum denoise strength is set by 'Inpaint denoising strength'. 0 = disabled; 1 = linear; 2-4 = recommended"
+        ),
+    )
+
+    shared.opts.add_option(
+        "ad_match_inpaint_bbox_size",
+        shared.OptionInfo(
+            default=InpaintBBoxMatchMode.OFF.value,  # Off
+            component=gr.Radio,
+            component_args={"choices": INPAINT_BBOX_MATCH_MODES},
+            label="Try to match inpainting size to bounding box size, if 'Use separate width/height' is not set",
+            section=section,
+        ).info(
+            "Strict is for SDXL only, and matches exactly to trained SDXL resolutions. Free works with any model, but will use potentially unsupported dimensions."
         ),
     )
 
@@ -1135,7 +1272,8 @@ def make_axis_on_xyz_grid():
         return
 
     model_list = ["None", *model_mapping.keys()]
-    samplers = [sampler.name for sampler in all_samplers]
+    xyz_samplers = [sampler.name for sampler in all_samplers]
+    xyz_schedulers = [scheduler.label for scheduler in schedulers]
 
     axis = [
         xyz_grid.AxisOption(
@@ -1175,6 +1313,11 @@ def make_axis_on_xyz_grid():
             partial(set_value, field="ad_denoising_strength"),
         ),
         xyz_grid.AxisOption(
+            "[ADetailer] CFG scale 1st",
+            float,
+            partial(set_value, field="ad_cfg_scale"),
+        ),
+        xyz_grid.AxisOption(
             "[ADetailer] Inpaint only masked 1st",
             str,
             partial(set_value, field="ad_inpaint_only_masked"),
@@ -1189,7 +1332,18 @@ def make_axis_on_xyz_grid():
             "[ADetailer] ADetailer sampler 1st",
             str,
             partial(set_value, field="ad_sampler"),
-            choices=lambda: samplers,
+            choices=lambda: xyz_samplers,
+        ),
+        xyz_grid.AxisOption(
+            "[ADetailer] ADetailer scheduler 1st",
+            str,
+            partial(set_value, field="ad_scheduler"),
+            choices=lambda: xyz_schedulers,
+        ),
+        xyz_grid.AxisOption(
+            "[ADetailer] noise multiplier 1st",
+            float,
+            partial(set_value, field="ad_noise_multiplier"),
         ),
         xyz_grid.AxisOption(
             "[ADetailer] ControlNet model 1st",